+++ src/universalchess/epaper/help_dialog.py
@@ -50,8 +50,8 @@ class HelpDialogWidget(Widget):
     def _get_font_loader(self):
         if self._font_loader is None:
             from universalchess.resources import ResourceLoader
-            self._font_loader = ResourceLoader(
-                "/opt/universalchess/resources", "/home/pi/resources")
+            from universalchess.paths import RESOURCES_DIR, USER_RESOURCES_DIR
+            self._font_loader = ResourceLoader(RESOURCES_DIR, USER_RESOURCES_DIR)
         return self._font_loader
 
     def _wrap(self, draw: ImageDraw.ImageDraw, text: str, font, max_width: int):